+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 15.5.12
+* [Media Blocks SDK .Net] Added Intel QuickSync H264, HEVC, VP9 and MJPEG encoders support.
+* [Video Edit SDK .Net] Added FastEdit_ExtractAudioStreamAsync method to extract the audio stream from the video file.
+* [Video Edit SDK .Net] Added "Audio Extractor" WinForms sample.
+
 ## 15.5.11
 * [Media Blocks SDK .Net] Updated MP4SinkBlock. Sink can split output files by duration, file size or timecode. Use MP4SplitSinkSettings instead MP4SinkSettings to configure.
 * [Video Capture SDK .Net] Added the OnMJPEGLowLatencyRAWFrame event that fired when the MJPEG low latency engine received a RAW frame from a camera.